[AUS] expose metrics for addon upgrades
expose metrics for addon upgrades

* `aus_addon_upgrade_policy_info` - cluster and addon metadata, including upgrade policy details
* `aus_addon_version_remaining_soak_days` - remaining soak days for a addon version on a cluster

they have the same set of labels as their cluster upgrade counterparts but have an additional one called `addon` to distinguish between addons managed on the same cluster.

part of APPSRE-8524
